style: rename to servrer
