+++ server/src/main/java/SDD/Server/Notes/NoteNotFoundException.java
@@ -1,4 +1,4 @@
-package SDD.PuckNotes.Notes;
+package SDD.Server.Notes;
 
 public class NoteNotFoundException extends RuntimeException {
