+++ server/src/main/java/SDD/Server/Notes/NoteService.java
@@ -1,4 +1,4 @@
-package SDD.PuckNotes.Notes;
+package SDD.Server.Notes;
 
 import org.bson.BsonBinarySubType;
 import org.bson.types.Binary;
