+++ server/src/main/java/SDD/Server/ServerApplication.java
@@ -1,13 +1,13 @@
-package SDD.PuckNotes;
+package SDD.Server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class PuckNotesApplication {
+public class ServerApplication {
 
   public static void main(String[] args) {
-    SpringApplication.run(PuckNotesApplication.class, args);
+    SpringApplication.run(ServerApplication.class, args);
   }
 
 }
