+++ server/src/main/resources/application.example.properties
@@ -1,2 +1,2 @@
-spring.application.name=PuckNotes
+spring.application.name=Server
 spring.data.mongodb.uri="MONGO URI W/ DATABASE NAME"
